added render_nil to Representable::Declarative
which can be used to easily render all nil values instead of
having to explicitly say so for each property:

```ruby
Module.new do
  include Representable::JSON
  render_nil
  property :name
  property :groupies
end
```

The problem with a `default_options` method as discussed e.g.
[here](https://github.com/apotonick/representable/issues/43) is that there are
various methods that all take different options, thus it would be unclear which
default options would be applied to which method etc.